Conflicts:
	.pre-commit-config.yaml

+++ src/onegov/core/datamanager.py
@@ -7,10 +7,7 @@ from onegov.core.utils import safe_move
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    # NOTE: Technically this could be overwritten by anything that
-    #       satisfies the ITransaction interface, but we are happier
-    #       not having to deal with the zope.interface mypy plugin
-    from transaction import Transaction
+    from transaction.interfaces import ITransaction
 
 
 class FileDataManager:
@@ -19,39 +16,51 @@ class FileDataManager:
 
     transaction_manager = transaction.manager
 
-    def __init__(self, data: bytes, path: str):
+    def __init__(self, data: bytes, path: str, tmp_path: str | None = None):
         self.data = data
         self.path = path
+        self.tmp_path = tmp_path
 
     @classmethod
-    def write_file(cls, data: bytes, path: str) -> None:
-        transaction.get().join(cls(data, path))
+    def write_file(
+        cls,
+        data: bytes,
+        path: str,
+        tmp_path: str | None = None
+    ) -> None:
+        transaction.get().join(cls(data, path, tmp_path))
 
     def sortKey(self) -> str:
         return 'files'
 
-    def commit(self, transaction: 'Transaction') -> None:
+    def commit(self, transaction: 'ITransaction') -> None:
         with tempfile.NamedTemporaryFile(delete=False) as temp:
             self.tempfn = temp.name
             temp.write(self.data)
 
-    def abort(self, transaction: 'Transaction') -> None:
+    def abort(self, transaction: 'ITransaction') -> None:
         pass
 
-    def tpc_vote(self, transaction: 'Transaction') -> None:
+    def tpc_vote(self, transaction: 'ITransaction') -> None:
         if not os.path.exists(self.tempfn):
-            raise ValueError('%s doesnt exist' % self.tempfn)
+            raise ValueError(f'{self.tempfn} doesnt exist')
         if os.path.exists(self.path):
             raise ValueError('file already exists')
 
-    def tpc_abort(self, transaction: 'Transaction') -> None:
+    def tpc_abort(self, transaction: 'ITransaction') -> None:
+        # if another DataManager before us in the chain raises a retryable
+        # error before we get to commit, we still have to execute tpc_abort
+        # despite tempfn not existing yet.
+        if not hasattr(self, 'tempfn'):
+            return
+
         try:
             os.remove(self.tempfn)
         except OSError:
             pass
 
-    def tpc_begin(self, transaction: 'Transaction') -> None:
+    def tpc_begin(self, transaction: 'ITransaction') -> None:
         pass
 
-    def tpc_finish(self, transaction: 'Transaction') -> None:
-        safe_move(self.tempfn, self.path)
+    def tpc_finish(self, transaction: 'ITransaction') -> None:
+        safe_move(self.tempfn, self.path, tmp_dst=self.tmp_path)
